Refactor following new model
