+++ front/lib/actions/mcp_internal_actions/servers/agent_copilot_context/index.ts
@@ -1,0 +1,29 @@
+import type { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
+
+import { AGENT_COPILOT_CONTEXT_TOOL_NAME } from "@app/lib/actions/mcp_internal_actions/servers/agent_copilot_context/metadata";
+import { TOOLS } from "@app/lib/actions/mcp_internal_actions/servers/agent_copilot_context/tools";
+import { registerTool } from "@app/lib/actions/mcp_internal_actions/tool_definition";
+import { makeInternalMCPServer } from "@app/lib/actions/mcp_internal_actions/utils";
+import type { AgentLoopContextType } from "@app/lib/actions/types";
+import type { Authenticator } from "@app/lib/auth";
+
+function createServer(
+  auth: Authenticator,
+  agentLoopContext?: AgentLoopContextType
+): McpServer {
+  const server = makeInternalMCPServer("agent_copilot_context");
+
+  for (const tool of TOOLS) {
+    registerTool(
+      auth,
+      server,
+      agentLoopContext,
+      AGENT_COPILOT_CONTEXT_TOOL_NAME,
+      tool
+    );
+  }
+
+  return server;
+}
+
+export default createServer;
