+++ front/lib/actions/mcp_internal_actions/tool_definition.ts
@@ -24,6 +24,7 @@ export type ToolHandlerExtra = RequestHandlerExtra<
   ServerNotification
 > & {
   agentLoopContext?: AgentLoopContextType;
+  auth?: Authenticator;
 };
 
 export type ToolHandlerResult = Result<CallToolResult["content"], MCPError>;
@@ -62,7 +63,8 @@ export function registerTool(
     withToolLogging(
       auth,
       { toolNameForMonitoring: monitoringName, agentLoopContext },
-      (params, extra) => tool.handler(params, { ...extra, agentLoopContext })
+      (params, extra) =>
+        tool.handler(params, { ...extra, agentLoopContext, auth })
     )
   );
 }
